refactor: 拆分 tasks.py 为模块包 + TaskRegistry 替代 _active_tasks
- 将 806 行的 tasks.py 拆分为 5 个模块（每个 ≤200 行）:
  - tasks/constants.py — 常量、结构化输出模型、系统提示词
  - tasks/parsing.py — chunk_text、parse_llm_result、build_case_info
  - tasks/summary.py — generate_summary、build_detail_zip_sync
  - tasks/batch_runner.py — 主逻辑、入口点、单文件分析
  - tasks/__init__.py — 重新导出 Django Q2 入口点
- 创建 TaskRegistry 类管理活跃 asyncio 任务引用
- batch_service.py 改用 task_registry.cancel() 替代直接 import _active_tasks
